Auto-reconnection for session expiration
Look at the second part of http://wiki.apache.org/hadoop/ZooKeeper/FAQ#A2

Basically, either the client lib has to emit an event to the application
or has to try to manage the end of session.

With the option autoReconnect, the client lib will silently reconnect the
application to zookeeper.

In this case, the application has to listen to the 'connect' event to restart
the watchers.